Fix string formatting for 64-bit mode

+++ WebViewProxy/WebViewProxy.m
@@ -193,7 +193,7 @@ static NSPredicate* webViewProxyLoopDetection;
     [self respondWithData:data mimeType:mimeType];
 }
 - (NSString*)_contentLength:(NSData*)data {
-    return [NSString stringWithFormat:@"%d", data.length];
+    return [NSString stringWithFormat:@"%lu", (unsigned long)data.length];
 }
 #endif
